Changed type of accountBalance in users table from integer to float. Added more console logs.

+++ data/migrations/20190321161651_usersTable.js
@@ -53,7 +53,7 @@ exports.up = function(knex, Promise) {
         
         //user account balance
         tbl
-            .integer('accountBalance')
+            .float('accountBalance')
             .defaultTo(0)
         
         //user last 4 credit card digits
